requirements.txt
Jinja2
altair
numpy
pandas
tensorflow
tensorflow-probability

# test/development dependencies
absl-py
pylint>=2.6.0
pytest
statsmodels
